requirements.txt
pyiceberg[pyarrow,s3fs,glue]==0.7.1
boto3>=1.34.0
snowflake-connector-python>=3.7.0
pyarrow>=15.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
faker>=24.0.0
